fixes #234 Investigate enabling more verbose compiler warnings
We enabled verbose compiler warnings, and found a lot of issues.
Some of these were even real bugs.  As a bonus, we actually save
some initialization steps in the compat layer, and avoid passing
some variables we don't need.

+++ src/core/defs.h
@@ -27,7 +27,7 @@
 #endif
 
 // Returns the size of an array in elements. (Convenience.)
-#define NNI_NUM_ELEMENTS(x) (sizeof(x) / sizeof((x)[0]))
+#define NNI_NUM_ELEMENTS(x) ((unsigned) (sizeof(x) / sizeof((x)[0])))
 
 // These types are common but have names shared with user space.
 // Internal code should use these names when possible.
